\name{what.weights}
\alias{what.weights}
\title{
Table of selected weights  
}
\description{
    Produce table of which weights were selected for each scenario
}
\usage{
what.weights(x)
}
\arguments{
  \item{x}{
    \code{envindex.bound} object, as produced by \code{\link{envindex.diff.qp}}
}
}
\details{

}
\value{
data.frame with columns assetid,ctf,species,use.dur,diff and attribute values
}
\author{
Joseph Guillaume
}
\seealso{
  \code{\link{envindex.diff.qp}}
}
\examples{
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
