\name{what.weights}
\alias{what.weights}
\title{
Table of selected weights  
}
\description{
    Produce table of which weights were selected for each scenario
}
\usage{
what.weights(x)
}
\arguments{
  \item{x}{
    \code{envindex.bound} object, as produced by \code{\link{envindex.diff.qp}}
}
}
\details{

}
\value{
data.frame with columns assetid,ctf,species,use.dur,run.type,diff and values of attribute weights
}
\author{
Joseph Guillaume
}
\seealso{
  \code{\link{envindex.diff.qp}}
}
\examples{
}
